server: move generic layers into package
